Database upgrade for transfers
Database has been upgraded to version 9 and now includes a table for
transfer information.  The functionality to insert transfer information
to the database has also been implemented.  The layout for the transfer
activity is updated to match that of all other activities. Function
added to transfer, add_bill, add_expense, and addAccount to hide
EditText keyboard on touch.